+++ tests/scrapers/office_test.py
@@ -10,17 +10,17 @@ This module tests that:
       being recorded in scraper errors.
     - Scraper uses parallel instances of LibreOffice properly.
     - The following MIME type and version combinations are supported:
-        - application/vnd.oasis.opendocument.text, 1.1
+        - application/vnd.oasis.opendocument.text, 1.2
         - application/msword, 97-2003
         - application/vnd.openxmlformats-officedocument.wordprocessingml.document, 2007 onwards
-        - application/vnd.oasis.opendocument.presentation, 1.1
+        - application/vnd.oasis.opendocument.presentation, 1.2
         - application/vnd.ms-powerpoint, 97-2003
         - application/vnd.openxmlformats-officedocument.presentationml.presentation, 2007 onwards
-        - application/vnd.oasis.opendocument.spreadsheet, 1.1
+        - application/vnd.oasis.opendocument.spreadsheet, 1.2
         - application/vnd.ms-excel, 8X
         - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, 2007 onwards
-        - application/vnd.oasis.opendocument.graphics, 1.1
-        - application/vnd.oasis.opendocument.formula, 1.0
+        - application/vnd.oasis.opendocument.graphics, 1.2
+        - application/vnd.oasis.opendocument.formula, 1.2
     - These MIME types are also supported with a made up version or None as
       the version.
     - A made up MIME type is not supported.
@@ -43,22 +43,22 @@ BASEPATH = "tests/data"
 @pytest.mark.parametrize(
     ["filename", "mimetype"],
     [
-        ("valid_1.1.odt", "application/vnd.oasis.opendocument.text"),
+        ("valid_1.2.odt", "application/vnd.oasis.opendocument.text"),
         ("valid_97-2003.doc", "application/msword"),
         ("valid_2007 onwards.docx", "application/vnd.openxmlformats-"
          "officedocument.wordprocessingml.document"),
-        ("valid_1.1.odp",
+        ("valid_1.2.odp",
          "application/vnd.oasis.opendocument.presentation"),
         ("valid_97-2003.ppt", "application/vnd.ms-powerpoint"),
         ("valid_2007 onwards.pptx", "application/vnd.openxml"
          "formats-officedocument.presentationml.presentation"),
-        ("valid_1.1.ods",
+        ("valid_1.2.ods",
          "application/vnd.oasis.opendocument.spreadsheet"),
         ("valid_8X.xls", "application/vnd.ms-excel"),
         ("valid_2007 onwards.xlsx", "application/vnd."
          "openxmlformats-officedocument.spreadsheetml.sheet"),
-        ("valid_1.1.odg", "application/vnd.oasis.opendocument.graphics"),
-        ("valid_1.0.odf", "application/vnd.oasis.opendocument.formula"),
+        ("valid_1.2.odg", "application/vnd.oasis.opendocument.graphics"),
+        ("valid_1.2.odf", "application/vnd.oasis.opendocument.formula"),
     ]
 )
 def test_scraper_valid_file(filename, mimetype, evaluate_scraper):
@@ -82,27 +82,27 @@ def test_scraper_valid_file(filename, mimetype, evaluate_scraper):
 @pytest.mark.parametrize(
     ["filename", "mimetype", "application"],
     [
-        ("valid_1.1.odt", "application/vnd.oasis.opendocument.text",
+        ("valid_1.2.odt", "application/vnd.oasis.opendocument.text",
          "writer"),
         ("valid_97-2003.doc", "application/msword", "writer"),
         ("valid_2007 onwards.docx",
          "application/vnd.openxmlformats-officedocument.wordprocessingml.document", # noqa
          "writer"),
-        ("valid_1.1.odp",
+        ("valid_1.2.odp",
          "application/vnd.oasis.opendocument.presentation", "impress"),
         ("valid_97-2003.ppt", "application/vnd.ms-powerpoint", "impress"),
         ("valid_2007 onwards.pptx",
          "application/vnd.openxmlformats-officedocument.presentationml.presentation", # noqa
          "impress"),
-        ("valid_1.1.ods",
+        ("valid_1.2.ods",
          "application/vnd.oasis.opendocument.spreadsheet", "calc"),
         ("valid_8X.xls", "application/vnd.ms-excel", "calc"),
         ("valid_2007 onwards.xlsx",
          "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
          "calc"),
-        ("valid_1.1.odg", "application/vnd.oasis.opendocument.graphics",
+        ("valid_1.2.odg", "application/vnd.oasis.opendocument.graphics",
          "draw"),
-        ("valid_1.0.odf", "application/vnd.oasis.opendocument.formula",
+        ("valid_1.2.odf", "application/vnd.oasis.opendocument.formula",
          "math"),
     ]
 )
@@ -133,21 +133,21 @@ def test_scraper_correct_application(filename, mimetype, application):
 @pytest.mark.parametrize(
     ["filename", "mimetype"],
     [
-        ("invalid_1.1_corrupted.odt", "application/vnd.oasis.opendocument"
+        ("invalid_1.2_corrupted.odt", "application/vnd.oasis.opendocument"
          ".text"),
         ("invalid_2007 onwards_corrupted.docx", "application/"
          "vnd.openxmlformats-officedocument.wordprocessingml.document"),
-        ("invalid_1.1_corrupted.odp",
+        ("invalid_1.2_corrupted.odp",
          "application/vnd.oasis.opendocument.presentation"),
         ("invalid_2007 onwards_corrupted.pptx", "application/vnd.openxml"
          "formats-officedocument.presentationml.presentation"),
-        ("invalid_1.1_corrupted.ods",
+        ("invalid_1.2_corrupted.ods",
          "application/vnd.oasis.opendocument.spreadsheet"),
         ("invalid_2007 onwards_corrupted.xlsx", "application/vnd."
          "openxmlformats-officedocument.spreadsheetml.sheet"),
-        ("invalid_1.1_corrupted.odg", "application/vnd.oasis.opendocument"
+        ("invalid_1.2_corrupted.odg", "application/vnd.oasis.opendocument"
          ".graphics"),
-        ("invalid_1.0_corrupted.odf", "application/vnd.oasis.opendocument"
+        ("invalid_1.2_corrupted.odf", "application/vnd.oasis.opendocument"
          ".formula"),
     ]
 )
@@ -182,7 +182,7 @@ def _scrape(filename, mimetype):
 @pytest.mark.parametrize(
     ["filename", "mimetype"],
     [
-        ("valid_1.1.odt", "application/vnd.oasis.opendocument.text"),
+        ("valid_1.2.odt", "application/vnd.oasis.opendocument.text"),
     ]
 )
 def test_parallel_validation(filename, mimetype):
@@ -216,7 +216,7 @@ def test_office_returns_invalid_return_code():
     when the tool gives an invalid return code"""
     mimetype = "application/vnd.oasis.opendocument.text"
     path = os.path.join("tests/data", mimetype.replace("/", "_"))
-    testfile = os.path.join(path, "valid_1.1.odt")
+    testfile = os.path.join(path, "valid_1.2.odt")
 
     scraper = OfficeScraper(filename=testfile,
                             mimetype=mimetype)
@@ -229,20 +229,20 @@ def test_office_returns_invalid_return_code():
 @pytest.mark.parametrize(
     ["mime", "ver"],
     [
-        ("application/vnd.oasis.opendocument.text", "1.1"),
+        ("application/vnd.oasis.opendocument.text", "1.2"),
         ("application/msword", "97-2003"),
         ("application/vnd.openxmlformats-"
          "officedocument.wordprocessingml.document", "2007 onwards"),
-        ("application/vnd.oasis.opendocument.presentation", "1.1"),
+        ("application/vnd.oasis.opendocument.presentation", "1.2"),
         ("application/vnd.ms-powerpoint", "97-2003"),
         ("application/vnd.openxml"
          "formats-officedocument.presentationml.presentation", "2007 onwards"),
-        ("application/vnd.oasis.opendocument.spreadsheet", "1.1"),
+        ("application/vnd.oasis.opendocument.spreadsheet", "1.2"),
         ("application/vnd.ms-excel", "8X"),
         ("application/vnd."
          "openxmlformats-officedocument.spreadsheetml.sheet", "2007 onwards"),
-        ("application/vnd.oasis.opendocument.graphics", "1.1"),
-        ("application/vnd.oasis.opendocument.formula", "1.0"),
+        ("application/vnd.oasis.opendocument.graphics", "1.2"),
+        ("application/vnd.oasis.opendocument.formula", "1.2"),
     ]
 )
 def test_is_supported(mime, ver):